+++ multiqc/modules/sourmash/sourmash.py
@@ -8,12 +8,13 @@ from multiqc import config
 from multiqc.modules.base_module import BaseMultiqcModule
 
 from .compare import compare
+from .gather import gather
 
 # Initialise the logger
 log = logging.getLogger(__name__)
 
 
-class MultiqcModule(BaseMultiqcModule, compare):
+class MultiqcModule(BaseMultiqcModule, compare, gather):
     def __init__(self):
         super(MultiqcModule, self).__init__(
             name="sourmash",
@@ -27,6 +28,10 @@ class MultiqcModule(BaseMultiqcModule, compare):
         n["compare"] = self.parse_compare()
         if n["compare"] > 0:
             log.info("Found {} compare reports".format(n["compare"]))
+
+        n["gather"] = self.parse_gather()
+            if n["gather"] > 0:
+            log.info("Found {} gather reports".format(n["gather"]))
 
         if sum(n.values()) == 0:
             raise UserWarning